+++ tests/testthat/test-model.R
@@ -113,15 +113,19 @@ test_that("forecasts() returns a data.frame", {
     expect_equal(length(mock_calls(m)), 1)
     expect_equal(mock_args(m)[[1]][[2]], "http://example.com/api/model/5/forecasts/")
     expect_is(the_forecasts, "data.frame")
-    expect_equal(names(the_forecasts), c("id", "url", "forecast_model_url", "source", "timezero_url", "created_at",
-                                         "notes", "forecast_data_url"))
+    expect_equal(names(the_forecasts), c("id", "url", "forecast_model_url", "source", "timezero_url",
+                                         "timezero_timezero_date", "timezero_data_version_date",
+                                         "timezero_is_season_start", "created_at", "notes", "forecast_data_url"))
     expect_equal(nrow(the_forecasts), 2)  # 2 forecasts
-    expect_equal(ncol(the_forecasts), 8)
+    expect_equal(ncol(the_forecasts), 11)
 
     exp_row <- data.frame(id = 3, url = "http://example.com/api/forecast/3/",
                           forecast_model_url = "http://example.com/api/model/5/",
                           source = "docs-predictions.json",
                           timezero_url = "http://example.com/api/timezero/5/",
+                          timezero_timezero_date = as.Date("2011-10-02", YYYY_MM_DD_DATE_FORMAT),
+                          timezero_data_version_date = as.Date(NA),
+                          timezero_is_season_start = TRUE,
                           created_at = as.Date("2020-03-05T15:47:47.369231-05:00"),
                           notes = "a small prediction file",
                           forecast_data_url = "http://example.com/api/forecast/3/data/",
@@ -141,15 +145,19 @@ test_that("forecasts() can handle null notes", {
     expect_equal(length(mock_calls(m)), 1)
     expect_equal(mock_args(m)[[1]][[2]], "http://example.com/api/model/5/forecasts/")
     expect_is(the_forecasts, "data.frame")
-    expect_equal(names(the_forecasts), c("id", "url", "forecast_model_url", "source", "timezero_url", "created_at",
-                                         "notes", "forecast_data_url"))
+    expect_equal(names(the_forecasts), c("id", "url", "forecast_model_url", "source", "timezero_url",
+                                         "timezero_timezero_date", "timezero_data_version_date",
+                                         "timezero_is_season_start", "created_at", "notes", "forecast_data_url"))
     expect_equal(nrow(the_forecasts), 2)  # 2 forecasts
-    expect_equal(ncol(the_forecasts), 8)
+    expect_equal(ncol(the_forecasts), 11)
 
     exp_row <- data.frame(id = 3, url = "http://example.com/api/forecast/3/",
                           forecast_model_url = "http://example.com/api/model/5/",
                           source = "docs-predictions.json",
                           timezero_url = "http://example.com/api/timezero/5/",
+                          timezero_timezero_date = as.Date("2011-10-02", YYYY_MM_DD_DATE_FORMAT),
+                          timezero_data_version_date = as.Date(NA),
+                          timezero_is_season_start = TRUE,
                           created_at = as.Date("2020-03-05T15:47:47.369231-05:00"),
                           notes = as.character(NA),
                           forecast_data_url = "http://example.com/api/forecast/3/data/",